Integrate @dvsa/mes-driver-schema; Re-name lambdas based on terraform standard
