+++ .npmrc
@@ -1,1 +1,2 @@
 audit-level = high
+@dvsa:registry=https://npm.pkg.github.com
