+++ package.json
@@ -29,6 +29,7 @@
   },
   "dependencies": {
     "@dvsa/mes-microservice-common": "^0.8.1",
+    "@dvsa/mes-driver-schema": "^0.0.1",
     "aws-sdk": "^2.1200.0",
     "axios": "^0.27.2",
     "joi": "^17.6.2",
@@ -58,7 +59,7 @@
     "rimraf": "^3.0.2",
     "serverless": "^3.22.0",
     "serverless-dotenv-plugin": "^4.0.2",
-    "serverless-offline": "^9.2.6",
+    "serverless-offline": "^11.0.3",
     "serverless-webpack": "^5.9.0",
     "source-map-support": "^0.5.21",
     "supertest": "^6.2.4",
