+++ serverless.yml
@@ -9,20 +9,20 @@ package:
   individually: true
 
 functions:
-  getPhotograph:
-    handler: src/functions/getPhotograph/framework/handler.handler
+  getDriverPhotograph:
+    handler: src/functions/getDriverPhotograph/framework/handler.handler
     events:
       - http:
           path: 'driver/photograph/{drivingLicenceNumber?}'
           method: get
-  getSignature:
-    handler: src/functions/getSignature/framework/handler.handler
+  getDriverSignature:
+    handler: src/functions/getDriverSignature/framework/handler.handler
     events:
       - http:
           path: 'driver/signature/{drivingLicenceNumber?}'
           method: get
-  getStandardDriver:
-    handler: src/functions/getStandardDriver/framework/handler.handler
+  postStandardDriver:
+    handler: src/functions/postStandardDriver/framework/handler.handler
     events:
       - http:
           path: 'driver/standard'
