--- src/common/domain/driver-photograph.interface.ts
@@ -1,8 +1,0 @@
-interface Photograph {
-  image: string;
-  imageFormat: string;
-}
-
-export interface DriverPhotograph {
-  photograph: Photograph;
-}
