--- src/common/domain/driver-signature.interface.ts
@@ -1,8 +1,0 @@
-interface Signature {
-    image: string;
-    imageFormat: string;
-}
-
-export interface DriverSignature {
-    signature: Signature;
-}
