+++ src/functions/getDriverPhotograph/framework/handler.ts
